+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 0.9.1 (Unreleased)
+ENHANCEMENTS:
+* compute: use `min_disk_size` of image or `disk_size` of snapshot to set size of boot_disk on instance template for `yandex_compute_instance_group`.
+
 ## 0.9.0 (August 07, 2019)
 FEATURES:
 * **New Data Source:** `yandex_lb_network_load_balancer`
@@ -7,10 +10,9 @@ FEATURES:
 * **New Resource:** `yandex_lb_target_group`
 
 ENHANCEMENTS:
-* compute: use `min_disk_size` of image or 'disk_size' of snapshot to set size of boot_disk on instance create.
+* compute: use `min_disk_size` of image or `disk_size` of snapshot to set size of boot_disk on instance create.
 * compute: update instance resource spec and platform type in one request.
 
-## 0.8.2 (July 10, 2019)
 BUG FIXES:
 * compute: change attribute `folder_id` from Required to Optional for `yandex_compute_instance_group` resource [[#14](https://github.com/terraform-providers/terraform-provider-yandex/issues/14)].